requirements.txt
opencv-python==4.6.0.66
numpy==1.23.5
paddlepaddle==2.4.2
paddleocr==2.6.1.3
flask==2.2.3
python-dotenv==1.0.0
requests==2.28.2
openai==0.27.8
pillow==9.5.0
scikit-image==0.20.0
matplotlib==3.7.1 